+++ src/client/Config.ts
@@ -29,6 +29,18 @@ class Config {
     return this.vsConfig.get<number>('contextLines', DEFAULT_CONTEXT_LINES);
   }
 
+  getAllowMoveFile(): boolean {
+    return this.vsConfig.get<boolean>('allowMoveFile', false);
+  }
+
+  getAllowDeleteFile(): boolean {
+    return this.vsConfig.get<boolean>('allowDeleteFile', false);
+  }
+
+  getEnableCors(): boolean {
+    return this.vsConfig.get<boolean>('enableCors', false);
+  }
+
   getConfiguration() {
     return this.vsConfig;
   }